+++ src/slices/AppSlice.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk, createSelector, createSlice } from "@reduxjs/toolkit";
 import { ethers } from "ethers";
-import { Providers } from "src/helpers/providers/providers";
+import { Providers } from "src/helpers/providers/Providers";
 import { RootState } from "src/store";
 
 import { abi as sOHMv2 } from "../abi/sOhmv2.json";
